Add new formatter `plainwithrepo`
Some times we want to list the package information including the repoid
in the same line.

This patch is adding a new formatter `plainwithrepo` which provides both
the default package information and the repoid in the same line
separated by a space. This formater can also be called with `nevrr`
alias.

Examples:

```
$ fedrq pkgs yt-dlp -S -F plainwithrepo
yt-dlp-2023.01.06-2.fc38.noarch rawhide
```

```
$ fedrq pkgs yt-dlp -S -F nevrr
yt-dlp-2023.01.06-2.fc38.noarch rawhide
```