config: improve Release repository loading
Previously, a repository id that was defined in one of the repo files in
a release's `defpaths` and also defined in the system configuration
(when system_repos=true) would cause fedrq to crash when loading
repositories. Now, system repositories are loaded first and then
only repositories that are not already defined are loaded from defpaths.

system_repos=false works the same as before.