add new repository loading API and dep on requests
This works similarly to the Formatter API. Different RepoGs can be added
to a Repos container. There's a get_repo method to select a repository
from the container. Repos containers are immutable mappings, but they
support the `|` union operator to return a new Repos container.
Repositories may accept arguments. There's factory functions to create
RepoG classes based on a list repository specs or from a format string
that can be used when loading fedrq's toml configuration.

The `file` RepoG requires the requests library to download a .repo file
from a URL.

A private _copr_repo() method was added to fedrq.config.Release to parse
a string into a Copr URL for the `copr` RepoG. CoprRepoG also requires
requests to download a .repo file.

The ConfigError exception class was moved to a _config module to avoid
an import loop.